backend/requirements.txt
daphne
channels
websockets
twilio
python-dotenv
requests
django
groq
deepgram-sdk
elevenlabs
aiohttp
certifi
